Update Calculator.java
All the calculator functions are in a single class and method, in addition to its additional functionalities such as reset. So if we want to add other functionalities, this method and class will become extensive and difficult to analyze for your code compression. By implementing the Strategy pattern in the project, we can give the user or the system the option of choosing the best strategy for a given operation. We can also use the adapter pattern.